python_client/requirements.txt
# OPC UA客户端库
opcua==0.98.13

# 数据库相关
SQLAlchemy==2.0.23
pymysql==1.1.0

# 数据处理
pandas==2.1.4
numpy==1.26.2

# 日志和配置
python-dotenv==1.0.0

# 密码哈希
bcrypt==4.1.2
